install jest
