+++ .babelrc
@@ -1,0 +1,5 @@
+// .babelrc
+
+{
+  "presets": ["@babel/preset-env"],
+}
